Fixed: Siglum folding
A full text field with character folding was not specifically configured for sigla searches, which meant that any sigla with accented characters were not found when using only roman letters for searching.

Additionally a slight tweak to the probe query handler was made to make it behave closer to the "real" searches.

Fixes rism-digital/rism-online-issues#377