Fixed: Implement a case-sensitive folding field
This commit adds a case-sensitive folding text field configuration, for use with the siglum search. It will fold non-ASCII characters to Basic Latin, but it will not match on differences in case. This is specifically designed for the siglum search query field, where we want to match on any sigla that may vary by diacritics from the query, but not differ by case.